Refactor chatbot logic and add RAG functionality; update settings and import paths for improved structure and performance
