+++ .vscode/settings.json
@@ -1,3 +1,6 @@
 {
-  "git.ignoreLimitWarning": true
+  "git.ignoreLimitWarning": true,
+  "python.analysis.extraPaths": [
+    "./src/rag"
+  ]
 }
