+++ GUI/modules/chatbot.py
@@ -5,7 +5,7 @@ sys.path.append(os.path.abspath(os.path.join(os.path.dirname(__file__), "../../"
 
 import streamlit as st
 
-from .bot import (
+from ...src.chatbot.bot import (
     initialize_conversation,
     chatbot_conversation
 )
